Function signature and simple alternative
First, in Swift 3 and later, the naming convention is to not include
preposition in the method name, but rather make it the label name for
the parameter. Frankly, I would remove it entirely, and just call the
method `sum(_:)`, but I didn’t want to go too far.

Also, in Kotlin, you showed a simpler alternative calling an existing
standard function, so I thought we should show the equivalent
simplified example here.